+++ lib/scale.js
@@ -23,7 +23,12 @@ module.exports = function(opts) {
 			},
 			body: JSON.stringify(processProfiles)
 		})
-		.then(function(response) { return response.json(); })
+		.then(function(response) {
+			if (!response.ok) {
+				throw new Error("Failed to scale", response);
+			}
+			return response.json();
+		})
 		.then(function(scaleResults) {
 			logger.info("scaling complete");
 			return scaleResults;
